Add Clear Complete task functionality
