+++ src/modules/clearCompletedTasks.js
@@ -1,0 +1,8 @@
+import { displayTasks, todo } from './displaytasks.js';
+
+const clearCompletedTasks = () => {
+  todo.clearCompleted();
+  displayTasks();
+};
+
+export default clearCompletedTasks;
