+++ src/style.css
@@ -1,36 +1,117 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap');
 
-* {
-  font-family: 'Poppins', sans-serif;
-}
-
-ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
+body {
+  background-color: #5d5a58;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+  font-family: 'Roboto', sans-serif;
 }
 
 li {
-  list-style-type: none;
-  margin: 0;
+  list-style: none;
+}
+
+.heading {
+  display: flex;
+  align-items: center;
+  padding: 10px;
+  border-radius: 5px;
+  box-shadow: 0 2px 20px;
+  width: 305px;
+  justify-content: space-between;
+  background-color: antiquewhite;
+}
+
+.heading h1 {
+  font-weight: 900;
+  font-size: 20px;
+}
+
+.tasklist {
+  width: 325px;
+}
+
+.tasks {
+  list-style: none;
   padding: 0;
+  background-color: antiquewhite !important;
+}
+
+#taskchk {
+  display: flex;
+  gap: 5px;
+  justify-content: flex-start;
+}
+
+#taskdiv {
+  display: flex;
+  justify-content: space-between;
+}
+
+.item {
+  padding: 10px;
+  border-radius: 5px;
+  box-shadow: 0 2px 20px;
+  margin-bottom: 1px;
+  align-items: center;
+  justify-content: flex-start;
+  display: flex;
+  gap: 10%;
+  flex-direction: column;
+}
+
+.item label {
+  width: 325px;
+}
+
+.list-head {
+  display: flex;
+}
+
+.list-head input {
+  padding: 10px;
+  border-radius: 5px;
+  box-shadow: 0 2px 20px;
+  width: 305px;
+  border: none;
+}
+
+.list-head button {
+  position: absolute;
+  display: none;
+}
+
+.clear {
+  margin: 1rem 0 0 0;
+}
+
+.clear button {
+  border: none;
+  width: 325px;
+  height: 50px;
+  border-radius: 4px;
+  font-size: 15px;
+  box-shadow: 0 2px 20px;
+  color: black;
+}
+
+.itme-desc {
+  width: 90%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
 }
 
 .task-div {
   display: flex;
-  gap: 1rem;
+  gap: 6px;
 }
 
-.itme-desc {
-  display: flex;
-  align-items: center;
-}
-
-.itme-desc:nth-child(even) {
-  background-color: rgb(150, 150, 150);
-}
-
-.X {
-  /* align-self: flex-end; */
-  margin: 0 1rem 0 auto;
+#task-info {
+  background-color: #5d5a58;
+  color: white;
+  border: none;
+  margin: 1rem 0;
 }
